# docs/SERVER_MODE/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MrRSS is a modern, cross-platform desktop RSS reader with auto-translation, smart feed discovery, and AI-powered summarization.",
        "title": "MrRSS API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/WCY-dt/MrRSS",
            "email": "mail@ch3nyang.top"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.3.17"
    },
    "host": "localhost:1234",
    "basePath": "/api",
    "paths": {
...
        "/ai/usage/reset": {
            "post": {
                "description": "Reset the AI usage token counter to zero",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translation"
                ],
                "summary": "Reset AI usage counter",
                "responses": {
                    "200": {
                        "description": "Success status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/translate/article": {
            "post": {
                "description": "Translate an article's title to the target language (uses AI or Google based on settings)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translation"
                ],
                "summary": "Translate article title",
                "parameters": [
                    {
                        "description": "Translation request (article_id, title, target_language)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translation result (translated_title, limit_reached)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request (missing required fields)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/translate/text": {
            "post": {
                "description": "Translate any text to the target language (uses AI or Google based on settings)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translation"
                ],
                "summary": "Translate text",
                "parameters": [
                    {
                        "description": "Translation request (text, target_language)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translation result (translated_text, html)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (missing required fields)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/translation/test-custom": {
            "post": {
                "description": "Test a custom translation API configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translation"
                ],
                "summary": "Test custom translation",
                "parameters": [
                    {
                        "description": "Test request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/translation.TestCustomTranslationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test result",
                        "schema": {
                            "$ref": "#/definitions/translation.TestCustomTranslationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/translations/clear": {
            "post": {
                "description": "Clear all translated article titles from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translation"
                ],
                "summary": "Clear all translations",
                "responses": {
                    "200": {
                        "description": "Success status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "translation.CustomTranslatorConfig": {
            "type": "object",
            "properties": {
                "bodyTemplate": {
                    "description": "Request body template with placeholders",
                    "type": "string"
                },
                "endpoint": {
                    "description": "API endpoint URL",
                    "type": "string"
                },
                "headers": {
                    "description": "Custom HTTP headers",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "langCodeMapping": {
                    "description": "Custom language code mapping",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "method": {
                    "description": "HTTP method (GET/POST)",
                    "type": "string"
                },
                "name": {
                    "description": "Configuration name",
                    "type": "string"
                },
                "responsePath": {
                    "description": "JSONPath to extract translation from response",
                    "type": "string"
                },
                "timeout": {
                    "description": "Request timeout in seconds",
                    "type": "integer"
                }
            }
        },
        "translation.TestCustomTranslationRequest": {
            "type": "object",
            "properties": {
                "config": {
                    "$ref": "#/definitions/translation.CustomTranslatorConfig"
                },
                "target_lang": {
                    "type": "string",
                    "example": "zh"
                },
                "text": {
                    "type": "string",
                    "example": "Hello, world!"
                }
            }
        },
        "translation.TestCustomTranslationResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}